Refactored, and improved code.

+++ ModAPI/Vehicle.cpp
@@ -7,4 +7,4 @@ ModAPI::Vehicle::Vehicle(const VehicleHandle handle) : Entity(handle)
 {}
 
 bool ModAPI::Vehicle::IsSeatFree(const eVehicleSeat seat) const
-{ return VEHICLE::IS_VEHICLE_SEAT_FREE(handle, static_cast<int>(seat), false); }
+{ return VEHICLE::IS_VEHICLE_SEAT_FREE(handle, seat, false); }
